repositories refactoring following entities changes

+++ src/main/java/com/rmr/dinosaurs/infrastucture/database/CourseInfoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourseInfoRepository extends JpaRepository<CourseInfo, Integer> {
+public interface CourseInfoRepository extends JpaRepository<CourseInfo, Long> {
 
 }
